Move spatial disaggregation to new submodule disaggregate
