+++ etrago/disaggregate/__init__.py
@@ -1,0 +1,7 @@
+"""
+
+"""
+
+__copyright__ = "tba"
+__license__ = "tba"
+__author__ = "tba"
