+++ etrago/disaggregate/spatial.py
