+++ etrago/tools/network.py
@@ -33,7 +33,7 @@ if "READTHEDOCS" not in os.environ:
     from etrago.tools import db
 
 from etrago import __version__
-from etrago.cluster.disaggregation import run_disaggregation
+from etrago.disaggregate.spatial import run_disaggregation
 from etrago.cluster.electrical import ehv_clustering, run_spatial_clustering
 from etrago.cluster.gas import run_spatial_clustering_gas
 from etrago.cluster.snapshot import skip_snapshots, snapshot_clustering
